+++ testing/test_junitxml.py
@@ -616,6 +616,9 @@ def test_dont_configure_on_slaves(tmpdir):
             self.pluginmanager = self
             self.option = self
 
+        def getini(self, name):
+            return "pytest"
+
         junitprefix = None
         # XXX: shouldnt need tmpdir ?
         xmlpath = str(tmpdir.join('junix.xml'))
@@ -1032,3 +1035,26 @@ def test_url_property(testdir):
     test_case = minidom.parse(str(path)).getElementsByTagName('testcase')[0]
 
     assert (test_case.getAttribute('url') == test_url), "The URL did not get written to the xml"
+
+
+@pytest.mark.parametrize('suite_name', ['my_suite', ''])
+def test_set_suite_name(testdir, suite_name):
+    if suite_name:
+        testdir.makeini("""
+            [pytest]
+            junit_suite_name={0}
+        """.format(suite_name))
+        expected = suite_name
+    else:
+        expected = 'pytest'
+    testdir.makepyfile("""
+        import pytest
+
+        def test_func():
+            pass
+    """)
+    result, dom = runandparse(testdir)
+    assert result.ret == 0
+    node = dom.find_first_by_tag("testsuite")
+    node.assert_attr(name=expected)
+